[Server] / #53 / adminpage accessToken 및 refreshToken 유효성 검사 전체 추가

+++ server/controllers/admin/chef.js
@@ -1,10 +1,17 @@
 const { review, chef, user, course, reservation } = require('../../models');
+const { isAuthorized, basicAccessToken } = require('../token/accessToken');
+const { refreshAuthorized } = require('../token/refreshToken');
 
 module.exports = {
   get: async (req, res) => {
     const cuisine = decodeURI(decodeURIComponent(req.params.cuisine));
     const reqCuisine = req.params.cuisine;
     const sendChefInfoArr = [];
+
+    const accessVerify = isAuthorized(req);
+
+    const startSlice = Number(req.query.startNum);
+    const endSlice = Number(req.query.endNum);
 
     const findChef = await chef.findAll({ where: { cuisine: cuisine } });
 
@@ -19,17 +26,35 @@ module.exports = {
       sendChefInfoArr.push(findChef[i].dataValues);
     }
 
-    const startSlice = Number(req.query.startNum);
-    const endSlice = Number(req.query.endNum);
+    if (!accessVerify) {
+      const refreshVerify = refreshAuthorized(req);
+      if (!refreshVerify) {
+        res.status(401).json({ message: 'Send new Login Request' });
+      } else {
+        delete refreshVerify.exp;
+        const accessToken = basicAccessToken(refreshVerify);
 
-    if (req.query.startNum && req.query.endNum) {
-      res.status(200).json({
-        message: 'ok',
-        length: sendChefInfoArr.length,
-        data: sendChefInfoArr.slice(startSlice, endSlice + 1),
-      });
-    } else if (!req.query.startNum || !req.query.endNum) {
-      res.status(400).json({ message: 'undefined chef' });
+        if (req.query.startNum && req.query.endNum) {
+          res.status(201).json({
+            accessToken,
+            message: 'ok',
+            length: sendChefInfoArr.length,
+            data: sendChefInfoArr.slice(startSlice, endSlice + 1),
+          });
+        } else if (!req.query.startNum || !req.query.endNum) {
+          res.status(400).json({ message: 'undefined chef' });
+        }
+      }
+    } else {
+      if (req.query.startNum && req.query.endNum) {
+        res.status(200).json({
+          message: 'ok',
+          length: sendChefInfoArr.length,
+          data: sendChefInfoArr.slice(startSlice, endSlice + 1),
+        });
+      } else if (!req.query.startNum || !req.query.endNum) {
+        res.status(400).json({ message: 'undefined chef' });
+      }
     }
   },
 
@@ -38,15 +63,38 @@ module.exports = {
     const findChefId = await chef.findOne({ id: req.body.id });
     const userId = findChefId.dataValues.chUserId;
 
-    await review.destroy({ where: { rvChefId: req.body.id }, force: true });
-    await reservation.destroy({
-      where: { rsChefId: req.body.id },
-      force: true,
-    });
-    await course.destroy({ where: { coChefId: req.body.id }, force: true });
-    await chef.destroy({ where: { id: req.body.id }, force: true });
-    await user.destroy({ where: { id: userId }, force: true });
+    const accessVerify = isAuthorized(req);
 
-    res.status(200).json({ message: 'ok' });
+    if (!accessVerify) {
+      const refreshVerify = refreshAuthorized(req);
+      if (!refreshVerify) {
+        res.status(401).json({ message: 'Send new Login Request' });
+      } else {
+        delete refreshVerify.exp;
+        const accessToken = basicAccessToken(refreshVerify);
+
+        await review.destroy({ where: { rvChefId: req.body.id }, force: true });
+        await reservation.destroy({
+          where: { rsChefId: req.body.id },
+          force: true,
+        });
+        await course.destroy({ where: { coChefId: req.body.id }, force: true });
+        await chef.destroy({ where: { id: req.body.id }, force: true });
+        await user.destroy({ where: { id: userId }, force: true });
+
+        res.status(201).json({ accessToken, message: 'ok' });
+      }
+    } else {
+      await review.destroy({ where: { rvChefId: req.body.id }, force: true });
+      await reservation.destroy({
+        where: { rsChefId: req.body.id },
+        force: true,
+      });
+      await course.destroy({ where: { coChefId: req.body.id }, force: true });
+      await chef.destroy({ where: { id: req.body.id }, force: true });
+      await user.destroy({ where: { id: userId }, force: true });
+
+      res.status(200).json({ message: 'ok' });
+    }
   },
 };
